wgengine/magicsock: isolate endpoint migration regression test
